refactor(codegen): split function.rs and linker.rs under the 2,000-line cap
Pure file split, no behaviour change.
